requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.4
